Now charging the user and saving replies etc.

+++ src/main/kotlin/net/bosowski/chattergpt/data/models/authentication/OauthUser.kt
@@ -37,7 +37,7 @@ class OauthUser : OidcUser {
     var username: String? = null
 
     @NotNull
-    var availableCredits: Float = 0f
+    var availableCredits: Double = 0.0
 
     @OneToOne(cascade = [CascadeType.ALL])
     @JoinColumn(name = "oauth_token_id")
